chore: add debug script
