+++ package.json
@@ -4,7 +4,6 @@
   "version": "2.9.1",
   "description": "Build documentation using TypeDoc",
   "type": "module",
-  "main": "index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/zakodium/typedoc-action.git"
